Move ComparisonPassed functions from WCSimRootEvent & WCSimRootGeom to a new (classless) file: WCSimRootTools. Don't use templates in WCSim tools because I can't figure out how to correctly linkdef them (hopefully someone else can)
